requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
curlify==2.2.1
facebook-business==5.0.4
facebookads==2.11.4
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.1.0
httplib2==0.16.0
idna==2.8
oauth2client==4.1.3
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.10.5
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
selenium==3.141.0
six==1.14.0
SQLAlchemy==1.3.13
tabulate==0.8.6
uritemplate==3.0.1
urllib3==1.25.7
worksheet==0.0.1a0
xdg==4.0.1
